+++ src/collective/contact/core/browser/organization.py
@@ -35,6 +35,7 @@ class Organization(BaseView):
     parent_organizations = []
     sub_organizations = []
     positions = []
+    activity = ''
 
     def update(self):
         super(Organization, self).update()
@@ -45,6 +46,7 @@ class Organization(BaseView):
         factory = getUtility(IVocabularyFactory, "OrganizationTypesOrLevels")
         vocabulary = factory(self.context)
         self.type = vocabulary.getTerm(organization.organization_type).title
+        self.activity = self.context.activity
 
         contactable = IContactable(organization)
         organizations = contactable.organizations
@@ -63,7 +65,9 @@ class Organization(BaseView):
         self.email = contact_details['email']
         self.phone = contact_details['phone']
         self.cell_phone = contact_details['cell_phone']
+        self.fax = contact_details['fax']
         self.im_handle = contact_details['im_handle']
+        self.website = contact_details['website']
         self.address = contact_details['address']
 
         # also show fields that were added TTW